backend/pyproject.toml
[project]
name = "sigmamatch-backend"
version = "0.1.0"
description = "SigmaMatch – backend matching engine and API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "pydantic>=2,<3",
    "pyyaml>=6,<7",
    "python-multipart>=0.0.9,<1",
    "slowapi>=0.1.9,<1",
    "defusedxml>=0.7,<1",
    "redis>=5,<6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "httpx>=0.27,<1",
    "ruff>=0.5,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100
